Revert "Automate ncoa (#134)"
This reverts commit d8bded5852e6e62b2296d165cdab507295760d3f.
