+++ .github/workflows/django.yml
@@ -4,7 +4,7 @@ on: [push, pull_request]
   # push:
   #   branches-ignore:
   #     # ignore all except master
-  #     - '*' # 'master'
+  #     - '*' # 'master'      
   # pull_request:
   #   branches:
   #     - '*'
@@ -35,7 +35,7 @@ jobs:
     strategy:
       max-parallel: 4
       matrix:
-        python-version: ['3.8', '3.9', '3.10']
+        python-version: [3.8, 3.9]
 
     steps:
       - uses: actions/checkout@v3
@@ -68,33 +68,26 @@ jobs:
         run: |
           cd backend
           python manage.py makemigrations
-          python manage.py migrate
-      - name: Run Tests and Code Coverage
+          python manage.py migrate      
+      - name: Run Tests and Code Coverage        
         run: |
           cd backend
           coverage run -m pytest
-          coverage report -m
-      - name: Start Server
-        if: matrix.python-version == '3.10'
-        run: |
-          cd backend
-          python manage.py collectstatic --noinput
-          python manage.py runserver &
-      - name: Wait for backend server to start
-        if: matrix.python-version == '3.10'
-        run: sleep 10
-      - name: Check Backend Server
-        if: matrix.python-version == '3.10'
-        run: |
-          echo "Checking if backend server is running..."
-          sleep 5 # Give the server some time to start
-          response=$(curl -s -o /dev/null -w "%{http_code}" http://localhost:8000/api/v1/accounts/) # Replace the URL with the appropriate endpoint of your backend
-          if [ "$response" != "200" ]; then
-            echo "Backend server is not running. Exiting. Code was $response"
-            exit 1
-          else
-            echo "Backend server is running."
-          fi
+          coverage report -m         
+          
+      # - name: Run Integration Tests
+      #   run: |
+      #     cd backend
+      #     # Add commands to run your integration tests
+
+      # - name: Deployment
+      #   run: |
+      #     # Add commands to deploy your application to the production environment
+      #     # If everything passes, heroku is notified and the application is deployed
+      # - name: Notify Status
+      #   run: |
+      #     # Add commands to send notifications about the CI/CD pipeline status
+      #     # Heroku already sends a notification if the deployment fails
 
   backend-security-check:
     needs: build-backend
@@ -105,7 +98,7 @@ jobs:
       - name: Set up Python
         uses: actions/setup-python@v3
         with:
-          python-version: '3.10'
+          python-version: 3.9
       - name: Install Dependencies
         run: |
           python -m pip install --upgrade pip
@@ -142,27 +135,7 @@ jobs:
           cd frontend
           yarn build
 
-  test-frontend:
-    runs-on: ubuntu-latest
-    needs: [build-backend, build-frontend]
-
-    steps:
-        - name: Checkout code
-          uses: actions/checkout@v2
-
-        - name: Set up Node.js
-          uses: actions/setup-node@v2
-          with:
-            node-version: 18
-        - name: Install dependencies
-          run: |
-            cd frontend
-            yarn install --frozen-lockfile
-        - name: Run tests
-          run: |
-            cd frontend
-            yarn test --passWithNoTests
-        # - name: Stop backend server
-        #   run: |
-        #     cd ../backend
-        #     python manage.py runserver --stop-graceful
+      - name: Run tests
+        run: |
+          cd frontend
+          yarn test --passWithNoTests
