+++ backend/config/settings.py
@@ -133,7 +133,7 @@ ROOT_URLCONF = "config.urls"
 CORS_ALLOWED_ORIGINS = [
     CLIENT_ORIGIN_URL,
     CLIENT_FRONEND_URL,
-    "https://is-my-customer-moving-git-develop-reidmhac.vercel.app",
+    "http://localhost:3000",
     # STRIPE
     "https://3.18.12.63",
     "https://3.130.192.231",
