+++ backend/data/management/commands/resyncClientList.py
@@ -2,7 +2,7 @@ from django.core.management.base import BaseCommand
 from datetime import datetime
 from accounts.models import Company
 from payments.models import Product
-from data.syncClients import complete_service_titan_sync
+from data.syncClients import get_service_titan_clients
 
 
 class Command(BaseCommand):
@@ -14,7 +14,7 @@ class Command(BaseCommand):
     def handle(self, *args, **options):
         company = options["company"]
         if company:
-            complete_service_titan_sync.delay(company_id=company, task_id=None)
+            get_service_titan_clients.delay(company_id=company, task_id=None)
         else:
             days_to_run = [0, 1, 2, 3, 4]
             current_datetime = datetime.now()
@@ -27,6 +27,6 @@ class Command(BaseCommand):
                     product=free_plan
                 )
                 for company in companies:
-                    complete_service_titan_sync.delay(
+                    get_service_titan_clients.delay(
                         company.id, task_id=None, automated=True
                     )
