+++ backend/payments/admin.py
@@ -1,15 +1,10 @@
 from django.contrib import admin
-from .models import Product, ServiceTitanInvoice
+from .models import Product
 
 
 class ProductAdmin(admin.ModelAdmin):
     list_display = ['id', 'name', 'amount', 'interval']
 
 
-class ServiceTitanInvoiceAdmin(admin.ModelAdmin):
-    list_display = ['id', 'amount', 'client', 'created_on']
-
-
 # Register your models here.
 admin.site.register(Product, ProductAdmin)
-admin.site.register(ServiceTitanInvoice, ServiceTitanInvoiceAdmin)
