+++ backend/requirements.txt
@@ -9,7 +9,7 @@ backoff==2.2.1
 beautifulsoup4==4.12.0
 billiard==3.6.4.0
 bs4==0.0.1
-cachetools==5.3.1
+cachetools==5.3.0
 celery==5.2.7
 certifi==2022.12.7
 cffi==1.15.0
@@ -61,16 +61,15 @@ lxml==4.9.2
 MarkupSafe==2.1.2
 oauthlib==3.2.2
 packaging==23.0
-pandas==2.0.3
 parsel==1.7.0
 platformdirs==3.1.1
-pluggy==1.2.0
+pluggy==1.0.0
 prompt-toolkit==3.0.38
 Protego==0.2.1
 psycopg2==2.9.5
 pyasn1==0.4.8
 pyasn1-modules==0.2.8
-pycparser==2.21
+pycparser==2.20
 PyDispatcher==2.0.7
 PyJWT==2.2.0
 pyOpenSSL==23.1.1
@@ -100,7 +99,7 @@ stripe==4.2.0
 tldextract==3.4.0
 tomli==2.0.1
 Twisted==22.10.0
-typing-extensions==4.7.1
+typing-extensions==3.10.0.2
 uritemplate==4.1.1
 urllib3==1.26.15
 vine==5.0.0
