+++ frontend/package.json
@@ -51,25 +51,29 @@
     "@react-google-maps/api": "^2.18.1",
     "@react-oauth/google": "^0.11.0",
     "@reduxjs/toolkit": "^1.9.1",
+    "apexcharts": "^3.35.0",
     "axios": "^0.26.1",
     "change-case": "^4.1.2",
-    "chart.js": "^4.3.0",
     "date-fns": "^2.28.0",
+    "dotenv": "^16.3.1",
     "file-saver": "^2.0.5",
     "font-awesome": "^4.7.0",
     "formik": "^2.4.2",
+    "history": "^5.2.0",
     "lodash": "^4.17.21",
     "numeral": "^2.0.6",
     "papaparse": "^5.3.2",
     "prop-types": "^15.8.1",
     "qrcode": "^1.5.1",
+    "qrcode.react": "^3.1.0",
     "react": "^17.0.2",
-    "react-chartjs-2": "^5.2.0",
+    "react-apexcharts": "^1.4.0",
     "react-countup": "^6.3.2",
     "react-dom": "^17.0.2",
     "react-google-button": "^0.7.2",
     "react-helmet-async": "^1.3.0",
     "react-redux": "^7.2.8",
+    "react-router": "6.11.2",
     "react-router-dom": "6.11.2",
     "react-scripts": "^5.0.0",
     "redux": "^4.1.2",
@@ -85,16 +89,19 @@
     "@babel/eslint-parser": "^7.17.0",
     "@testing-library/jest-dom": "^5.16.5",
     "@testing-library/react": "^12.1.5",
+    "@testing-library/user-event": "^14.4.3",
     "eslint": "^8.42.0",
     "eslint-config-airbnb": "^19.0.4",
     "eslint-config-prettier": "^8.4.0",
     "eslint-config-react-app": "^7.0.0",
+    "eslint-plugin-flowtype": "^8.0.3",
     "eslint-plugin-import": "^2.25.4",
     "eslint-plugin-jsx-a11y": "^6.7.1",
     "eslint-plugin-prettier": "^4.0.0",
     "eslint-plugin-react": "^7.28.0",
     "eslint-plugin-react-hooks": "^4.3.0",
     "prettier": "^2.5.1",
+    "redux-devtools-extension": "^2.13.9",
     "redux-mock-store": "^1.5.4"
   }
 }
