+++ frontend/src/components/Page.js
@@ -10,13 +10,7 @@ const Page = forwardRef(({ children, userInfo = null, title = '', meta, ...other
   <>
     <Helmet>
       <title>{`${title} | Did My Customer Move`}</title>
-      <link rel="apple-touch-icon" sizes="180x180" href="/apple-touch-icon.png" />
-      <link rel="icon" type="image/png" sizes="32x32" href="/favicon-32x32.png" />
-      <link rel="icon" type="image/png" sizes="16x16" href="/favicon-16x16.png" />
-      <link rel="manifest" href="/site.webmanifest" />
-      <link rel="mask-icon" href="/safari-pinned-tab.svg" color="#5bbad5" />
-      <meta name="msapplication-TileColor" content="#da532c" />
-      <meta name="theme-color" content="#ffffff" />
+      <link rel="icon" href="/favicon.ico" />
       {meta}
     </Helmet>
