--- frontend/src/components/charts/Bar.js
@@ -1,71 +1,0 @@
-import React from 'react';
-import { Bar } from 'react-chartjs-2';
-import { Chart, CategoryScale, LinearScale, BarElement, Title, Tooltip, Legend } from 'chart.js';
-import PropTypes from 'prop-types';
-
-import homeStyle from '../../theme/Home.module.css';
-
-Chart.register(CategoryScale, LinearScale, BarElement, Title, Tooltip, Legend);
-
-const BarChart = ({ title, values, dataLabel }) => {
-  const labels = ['January', 'February', 'March', 'April', 'May', 'June', 'July'];
-  const data = {
-    labels,
-    datasets: [
-      {
-        label: dataLabel,
-
-        data: values,
-        backgroundColor: [
-          'rgba(255, 99, 132, 0.2)',
-          'rgba(255, 159, 64, 0.2)',
-          'rgba(255, 205, 86, 0.2)',
-          'rgba(75, 192, 192, 0.2)',
-          'rgba(54, 162, 235, 0.2)',
-          'rgba(153, 102, 255, 0.2)',
-          'rgba(201, 203, 207, 0.2)',
-        ],
-        borderColor: [
-          'rgb(255, 99, 132)',
-          'rgb(255, 159, 64)',
-          'rgb(255, 205, 86)',
-          'rgb(75, 192, 192)',
-          'rgb(54, 162, 235)',
-          'rgb(153, 102, 255)',
-          'rgb(201, 203, 207)',
-        ],
-        borderWidth: 1,
-      },
-    ],
-  };
-  const options = {
-    responsive: true,
-    plugins: {
-      legend: {
-        display: false,
-      },
-      title: {
-        display: false,
-        text: 'Attributed Revenue',
-        font: {
-          size: 20,
-        },
-      },
-    },
-    maintainAspectRatio: false,
-  };
-  return (
-    <div className={homeStyle.main}>
-      <h2>{title}</h2>
-      <Bar data={data} options={options} />
-    </div>
-  );
-};
-
-BarChart.propTypes = {
-  title: PropTypes.string.isRequired,
-  values: PropTypes.arrayOf(PropTypes.number).isRequired,
-  dataLabel: PropTypes.string.isRequired,
-};
-
-export default BarChart;
