+++ frontend/src/pages/dashboard/RecentlySoldData.js
@@ -35,9 +35,6 @@ import RecentlySoldListCall from '../../redux/calls/RecentlySoldListCall';
 import { selectRecentlySold, getRecentlySoldCSV } from '../../redux/actions/usersActions';
 import { showLoginInfo } from '../../redux/actions/authActions';
 import { makeDate } from '../../utils/makeDate';
-import { handleChangePage, handleChangeRowsPerPage, handleRequestSort } from '../../utils/dataTableFunctions';
-import { getComparator, applySortFilter } from '../../utils/filterFunctions';
-
 
 // ----------------------------------------------------------------------
 
@@ -51,6 +48,34 @@ const TABLE_HEAD = [
   { id: 'year_built', label: 'Year Built', alignRight: false },
 ];
 
+// ----------------------------------------------------------------------
+// change this to sort by status
+function descendingComparator(a, b, orderBy) {
+  if (b[orderBy] < a[orderBy]) {
+    return -1;
+  }
+  if (b[orderBy] > a[orderBy]) {
+    return 1;
+  }
+  return 0;
+}
+
+export function getComparator(order, orderBy) {
+  return order === 'desc'
+    ? (a, b) => descendingComparator(a, b, orderBy)
+    : (a, b) => -descendingComparator(a, b, orderBy);
+}
+
+export function applySortFilter(array, comparator) {
+  const stabilizedThis = array.map((el, index) => [el, index]);
+
+  stabilizedThis.sort((a, b) => {
+    const order = comparator(a[0], b[0]);
+    if (order !== 0) return order;
+    return a[1] - b[1];
+  });
+  return stabilizedThis.map((el) => el[0]);
+}
 
 export default function RecentlySoldData() {
   const dispatch = useDispatch();
@@ -82,6 +107,24 @@ export default function RecentlySoldData() {
     }
     setRecentlySoldLength(RECENTLYSOLDLIST.length);
   }, [RECENTLYSOLDLIST, recentlySoldLength]);
+
+  const handleRequestSort = (event, property) => {
+    const isAsc = orderBy === property && order === 'asc';
+    setOrder(isAsc ? 'desc' : 'asc');
+    setOrderBy(property);
+  };
+
+  const handleChangePage = (event, newPage) => {
+    // fetch new page if two away from needing to see new page
+    // if (((newPage + 2) * rowsPerPage) % 1000 === 0) {
+    //   dispatch(recentlySoldAsync(((newPage + 2) * rowsPerPage) / 1000 + 1));
+    // }
+    setPage(newPage);
+  };
+  const handleChangeRowsPerPage = (event) => {
+    setRowsPerPage(parseInt(event.target.value, 10));
+    setPage(0);
+  };
 
   const [minPrice, setMinPrice] = useState('');
   const [maxPrice, setMaxPrice] = useState('');
@@ -281,7 +324,7 @@ export default function RecentlySoldData() {
                         headLabel={TABLE_HEAD}
                         rowCount={rowsPerPage}
                         numSelected={0}
-                        onRequestSort={(event, property) => handleRequestSort(event, property, orderBy, order, setOrder, setOrderBy)}
+                        onRequestSort={handleRequestSort}
                         onSelectAllClick={handleRequestSort}
                         checkbox={0}
                       />
@@ -358,17 +401,13 @@ export default function RecentlySoldData() {
                 </Box>
               )}
               <TablePagination
-                  rowsPerPageOptions={[10, 50, 100]}
-                  component="div"
-                  count={shownClients}
-                  rowsPerPage={rowsPerPage}
-                  page={page}
-                  onPageChange={(event, newPage) =>
-                      handleChangePage(event, newPage, setPage)
-                  }
-                  onRowsPerPageChange={(event) =>
-                      handleChangeRowsPerPage(event, setRowsPerPage, setPage)
-                  }
+                rowsPerPageOptions={[10, 50, 100]}
+                component="div"
+                count={shownClients}
+                rowsPerPage={rowsPerPage}
+                page={page}
+                onPageChange={handleChangePage}
+                onRowsPerPageChange={handleChangeRowsPerPage}
               />
             </Card>
             {/* TODO */}
