+++ frontend/src/routes.js
@@ -3,8 +3,7 @@ import { Navigate, useRoutes } from 'react-router-dom';
 import DashboardLayout from './layouts/dashboard';
 import LogoOnlyLayout from './layouts/LogoOnlyLayout';
 //
-import Dashboard from './pages/dashboard/Home';
-import CustomerData from './pages/dashboard/CustomerData';
+import HomePage from './pages/dashboard/Home';
 import ForSaleData from './pages/dashboard/ForSaleData';
 import RecentlySoldData from './pages/dashboard/RecentlySoldData';
 import Login from './pages/root/Login';
@@ -47,8 +46,7 @@ export default function Router() {
         </ProtectedRoute>
       ),
       children: [
-        { path: '', element: <Dashboard /> },
-        { path: 'customers', element: <CustomerData /> },
+        { path: '', element: <HomePage /> },
         { path: 'settings/user', element: <ProfileSettings /> },
         { path: 'settings/enterprise', element: <EnterpriseSettings /> },
         { path: 'forsale', element: <ForSaleData /> },
