--- frontend/src/theme/Home.module.css
@@ -1,8 +1,0 @@
-.main {
-    max-height: 40vh;
-    margin: 2rem;
-    display: flex;
-    flex-direction: column;
-    justify-content: center;
-    align-items: center;
-}
